Add error data log in console
