+++ wormhole-console/src/main/java/com/zergclan/wormhole/console/api/contorller/log/ErrorDataLogController.java
@@ -15,15 +15,14 @@
  * limitations under the License.
  */
 
-package com.zergclan.wormhole.console.api.contorller;
+package com.zergclan.wormhole.console.api.contorller.log;
 
+import com.zergclan.wormhole.console.api.contorller.AbstractRestController;
 import com.zergclan.wormhole.console.api.vo.HttpResult;
 import com.zergclan.wormhole.console.api.vo.PageQuery;
-import com.zergclan.wormhole.console.application.domain.entity.ErrorDataLog;
-import com.zergclan.wormhole.console.application.service.ErrorDataLogService;
+import com.zergclan.wormhole.console.application.domain.metrics.ErrorDataLog;
+import com.zergclan.wormhole.console.application.service.log.ErrorDataLogService;
 import com.zergclan.wormhole.console.infra.repository.PageData;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -35,22 +34,11 @@ import javax.annotation.Resource;
  * Controller of {@link ErrorDataLog}.
  */
 @RestController
-@RequestMapping("/log/error/data")
+@RequestMapping("/metrics/error/data")
 public class ErrorDataLogController extends AbstractRestController {
     
     @Resource
     private ErrorDataLogService errorDataLogService;
-    
-    /**
-     * Get {@link ErrorDataLog} by id.
-     *
-     * @param id id
-     * @return {@link HttpResult}
-     */
-    @GetMapping("/{id}")
-    public HttpResult<ErrorDataLog> getById(@PathVariable(value = "id") final Integer id) {
-        return success(errorDataLogService.getById(id));
-    }
     
     /**
      * List {@link ErrorDataLog} by {@link PageQuery}.
