+++ wormhole-console/src/main/java/com/zergclan/wormhole/console/infra/repository/mapper/ErrorDataLogMapper.java
@@ -17,7 +17,7 @@
 
 package com.zergclan.wormhole.console.infra.repository.mapper;
 
-import com.zergclan.wormhole.console.application.domain.entity.ErrorDataLog;
+import com.zergclan.wormhole.console.application.domain.metrics.ErrorDataLog;
 import org.apache.ibatis.annotations.Mapper;
 
 /**
