+++ wormhole-console/src/main/resources/Mapper/ErrorDataLogMapper.xml
@@ -1,0 +1,129 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Licensed to the Apache Software Foundation (ASF) under one or more
+  ~ contributor license agreements.  See the NOTICE file distributed with
+  ~ this work for additional information regarding copyright ownership.
+  ~ The ASF licenses this file to You under the Apache License, Version 2.0
+  ~ (the "License"); you may not use this file except in compliance with
+  ~ the License.  You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.zergclan.wormhole.console.infra.repository.mapper.ErrorDataLogMapper">
+    <resultMap id="ErrorDataLogResultMap" type="com.zergclan.wormhole.console.application.domain.metrics.ErrorDataLog">
+        <result column="id" property="id" jdbcType="INTEGER"/>
+        <result column="task_identifier" property="taskIdentifier" jdbcType="VARCHAR"/>
+        <result column="plan_batch" property="planBatch" jdbcType="BIGINT"/>
+        <result column="task_batch" property="taskBatch" jdbcType="BIGINT"/>
+        <result column="code" property="code" jdbcType="VARCHAR"/>
+        <result column="message" property="message" jdbcType="VARCHAR"/>
+        <result column="owner_identifier" property="ownerIdentifier" jdbcType="VARCHAR"/>
+        <result column="data_json" property="dataJson" jdbcType="VARCHAR"/>
+        <result column="create_timestamp" property="createTimestamp" jdbcType="BIGINT"/>
+    </resultMap>
+
+    <sql id="table_name">
+        error_data_log
+    </sql>
+
+    <sql id="id">
+        id = #{id,jdbcType=INTEGER}
+    </sql>
+
+    <sql id="limit">
+        LIMIT #{start}, #{size}
+    </sql>
+
+    <sql id="insert_columns">
+        task_identifier, plan_batch, task_batch, code, message, owner_identifier, data_json, create_timestamp
+    </sql>
+
+    <sql id="insert_values">
+        #{task_identifier,jdbcType=VARCHAR}, #{plan_batch,jdbcType=BIGINT}, #{task_batch,jdbcType=BIGINT},
+        #{code,jdbcType=VARCHAR}, #{message,jdbcType=VARCHAR}, #{owner_identifier,jdbcType=BIGINT},
+        #{data_json,jdbcType=BIGINT}, #{create_timestamp,jdbcType=BIGINT}
+    </sql>
+
+    <sql id="insert_batch_values">
+        #{batch.task_identifier,jdbcType=VARCHAR}, #{batch.plan_batch,jdbcType=BIGINT}, #{batch.task_batch,jdbcType=BIGINT},
+        #{batch.code,jdbcType=VARCHAR}, #{batch.message,jdbcType=VARCHAR}, #{batch.owner_identifier,jdbcType=BIGINT},
+        #{batch.data_json,jdbcType=BIGINT}, #{batch.create_timestamp,jdbcType=BIGINT}
+    </sql>
+
+    <sql id="select_columns">
+        id, task_identifier, plan_batch, task_batch, code, message, owner_identifier, data_json, create_timestamp
+    </sql>
+
+    <sql id="select_count">
+        COUNT(id) AS total
+    </sql>
+
+    <sql id="where_fuzzy_query">
+        <if test="null != taskBatch">
+            OR task_batch = #{task_batch,jdbcType=BIGINT}
+        </if>
+        <if test="null != code">
+            OR code = #{username,jdbcType=VARCHAR}
+        </if>
+        <if test="null != message">
+            OR message LIKE CONCAT('%', #{message,jdbcType=VARCHAR}, '%')
+        </if>
+        <if test="null != ownerIdentifier">
+            OR owner_identifier LIKE CONCAT('%', #{owner_identifier,jdbcType=VARCHAR}, '%')
+        </if>
+    </sql>
+
+    <sql id="where_fuzzy_query_page">
+        <if test="null != query.taskBatch">
+            OR task_batch = #{query.task_batch,jdbcType=BIGINT}
+        </if>
+        <if test="null != query.code">
+            OR code = #{query.username,jdbcType=VARCHAR}
+        </if>
+        <if test="null != query.message">
+            OR message LIKE CONCAT('%', #{query.message,jdbcType=VARCHAR}, '%')
+        </if>
+        <if test="null != query.ownerIdentifier">
+            OR owner_identifier LIKE CONCAT('%', #{query.owner_identifier,jdbcType=VARCHAR}, '%')
+        </if>
+    </sql>
+
+    <insert id="insert" parameterType="com.zergclan.wormhole.console.application.domain.metrics.ErrorDataLog">
+        INSERT INTO <include refid="table_name"/> (<include refid="insert_columns"/>) VALUES (<include refid="insert_values"/>)
+    </insert>
+
+    <insert id="insertBatch" parameterType="java.util.Collection">
+        INSERT INTO <include refid="table_name"/> (<include refid="insert_columns"/>)
+        VALUES
+        <foreach collection="list" item="batch" separator=",">
+            (<include refid="insert_batch_values"/>)
+        </foreach>
+    </insert>
+
+    <delete id="deleteById" parameterType="java.lang.Integer">
+        DELETE FROM <include refid="table_name"/> WHERE <include refid="id"/>
+    </delete>
+
+    <select id="countByPage" parameterType="com.zergclan.wormhole.console.application.domain.metrics.ErrorDataLog" resultType="java.lang.Integer">
+        SELECT <include refid="select_count"/> FROM <include refid="table_name"/>
+        <where>
+            <include refid="where_fuzzy_query"/>
+        </where>
+    </select>
+
+    <select id="page" parameterType="com.zergclan.wormhole.console.api.vo.PageQuery" resultMap="ErrorDataLogResultMap">
+        SELECT <include refid="select_columns"/> FROM <include refid="table_name"/>
+        <where>
+            <include refid="where_fuzzy_query_page"/>
+        </where>
+        <include refid="limit"/>
+    </select>
+</mapper>
